Fix that built-in error message response not HTML escaped.
Stack traces and Object#inspect output lots of less-than signs
that need escaping.

There's probably a better way to do this. My first iteration called
CGI::escapeHTML and then I noticed that Rack includes a nice Utils module with
various escaping methods so I figured that would be a better fit for Sinatra.
Unfortunately, these methods don't seem to be exposed anywhere (that I could
find) so I extended the Rack::Utils module with itself so that messages could be
sent directly to the module.