Fix a few minor issues with, and document, the POST _method hack.
The original issue I ran into was that the code was not letting downcase method
names through on the _method param. So, if the post body had "_method=put"
instead of "_method=PUT", it would be processed as a POST. While there, I fixed
a few other cases, including the following:

  * Allow for verbs other than PUT/DELETE. OPTIONS and HEAD were added but
    additional verbs can be added on an app-by-app basis by cat'ing onto the
    POST_TUNNEL_METHODS_ALLOWED array.

  * Only look for _method param in POST body, not in query string. The code
    previously used Request#params, which includes query string and post body
    parameters.

This patch includes tests that verify each of the previously described
behaviors.