Sinatra::Application as a sandbox for the DSL
Moves DSL methods previously defined on (main) to Sinatra::Application
and begins documenting them. This unifies a bunch of method names and
will allow individual Application instances to be used as a sandbox
for DSL execution.

The Sinatra::Application::FORWARD_METHODS array is a list of method
names that should be available at the top-level. A constant was created
so that its easy to tell what methods are available from top-level when
reading the API documentation. When one of the methods is invoked on
(main), the message is forwarded to the default application
(Sinatra::application).

This is in preparation for some configuration related changes and
multi-app support.