Rack pipelines leading to app; 'use' DSL method.
Moved Rack middleware management into Sinatra::Application and
enabled support for assembling custom Rack middleware pipelines
before an app.

  require 'sinatra'
  use Rack::Lint
  use Rack::Authentication
  get '/' do
    "Hello, World"
  end

Assuming a default set of options, this would result in the
following pipeline leading to the application:

  CommonLogger -> Lint -> Authentication -> Application

(The Rack::CommonLogger middleware is created automatically when the
:logging option is enabled.)

Also worth noting is that the pipeline leading to the application is
reassembled on each request during reload in :development mode.