Fix #18 by wrapping reload and pipeline in mutex lock
This fixes the issue of random errors and backtraces being
dumped to STDERR in development when you hit the site too-fast.
The problem seemed to be that reloading was not occuring within
the Mutex lock so there's a chance that objects get into some
kind of weird half-state.

With this patch, the mutex lock is always enabled in development
and I've moved the calls to run_safely out of Application#dispatch
and into Application#call so that the entire reloading process and
Rack pipeline run exclusively.