make raise_errors option more reliable
The raise_errors option was implemented as part of
the default error handler making it worthless in development
mode or with a custom error handler. This moves the
implementation to the core dispatch method.

While here, clean up the Error class a bit, making it
use the options provided and pass the correct HTTP status
code.