simplify Static implementation; use 8K chunks instead of 4K
The Static event handler is now instantiated with a reference
back to the Sinatra app instead of using the global singleton.
While here, bump streaming buffer size from 4K to 8K -- word
around the campfire is that reading in 8K chunks results in a bit
less IO on most modern systems.