Don't allow :session option in body/params arg; misc refactoring
This changes the parent commit's :session option semantics just
slightly. The :session option must be passed as part of the options
Hash -- the body arg is assumed to be params or a POST/PUT body.
Also, the mapping from :session to HTTP_COOKIE has been switched to
'rack.session' - mapping it to HTTP_COOKIE doesn't make any sense.

While here, refactor the make_request method to make it more
obvious that we're really just building up the options Hash for
MockRequest.