Make gemspec conflict free and add anything that is needed to run the spec.
The specs do not pass with haml 2.x, therefore depend on 3.0 or later. However, hanna depends on haml 2.2.14.
Removed hanna to get tests passing. Add rake and rdoc as dependencies (which were dependencies of hanna, therefore not needed previously).
Depend on rack-test >= 0.5.6 (0.5.5 was a bogus release that keeps the tests from passing).