Support HTTPS with WEBrick by allowing to pass options to the rack handler in Sinatra::Application.run!
This patch allows a user to pass custom options to Rack handlers in
Sinatra::Application.run!. This effectively allows a user to enable SSL
in WEBrick (Thin and Mongrel do not support HTTPS). See README.rdoc for
a snippet on how to enable HTTPS with WEBrick.

This patch also includes a unit test and updated documentation in
README.rdoc and README.de.rdoc.