Replaced define_method/remove_method (in generate_method) with instance_exec() to simplify code.
Now that 1.8.6 is no longer supported we no longer need the define_method/remove_method trick and can make
use of Ruby's own mechanism for this purpose - instance_exec(). Lambda-like arity semantics are maintained
by wrapping the block in a proc that checks arity. As in the previous situation an exception is made for arity 0 blocks,
where normal block semantics hold.