Set the Rack::File's root to the app file's root
`Rack::File`'s API doesn't accept a `nil` for root path. Earlier, we
were assigning the path of the file manually in Sinatra by the following
steps:

    file = Rack::File.new(nil)
    file.path = path
    file.serving(env)

But the API of `Rack::File#serving` has been changed to take two
arguments: `request` and `path` of the file. `Rack::File` would then
interpret the file path based on the root path supplied to it's `#new`
method, and the `path` value passed into the `#serving` method.

For these reasons, we decided to use `Rack::File.call` method, but this
causes issues with how the root path is set as `Rack::File` since it
appends the route fragment to the root path to fetch the file. This has
two side-effects:

1. The static file that's being served should be present in the public
directory
2. The static file path under `public` directory should match the route
fragment, including the extension. For example:

    get '/file.txt' do
	  send_file 'file.txt'
	end

will fetch the file at `<app root>/public/file.txt`, whereas:

    get '/file' do
	  send_file 'file.txt'
	end

will try to fetch the file at `<app root>/public/file` and will fail
since there won't be a `file.txt` at that location.

* This PR also fixes the failing specs that are caused by this appending
  of `public` directory effect by simply setting the public directory to
  app file root by using `settings.app_file` method.