Remove RedCloth from development dependencies
As far as I can tell, we're not actually testing it in any way.

And since they removed JRuby and Windows support in the latest release,
our only option is to lock to an older version. /cc jgarber/redcloth#24

But since we don't test it anyways what's the difference. :shrug: