avoid overwriting params with falsey value
This behavior arises from the commit that mustermann was introduced.
see: https://github.com/sinatra/sinatra/commit/7a9ffccdafb71fef8d33c110d03ca0654dbb9316

That breaks backward compatibility with Sinatra v1.4.X and is not intended behavior.
This changes it so that it overwrites params only if there is a possible value as a block argument.

Fixes #1417