Drop activesupport dependency
ActiveSupport was being loaded to provide Object#try. It's a very large
dependency that can monkey patch many core classes, and although the
require in sinatra-contrib is targeted and doesn't load all of
ActiveSupport, once the gem is on the load path some other libraries
will opportunistically load more of it.

Object#try is only used twice, maybe it's ok to live with a manual check
for nil for now? sinatra-contrib currently requires ruby 2.2 or better -
once the minimum is lifted to 2.3, Object#try will be provided by ruby
core.