Use Kernel#caller_locations if available
Ruby 2.x includes caller_locations in Kernel. The implementation
is subtly different from what Sinatra used to provide, which leads
to compatibility issues when an external caller tries to use
caller_locations.

This changes the previous implementation to be closer to caller_locations
as implemented in Kernel on modern Rubies, and delegates to caller_locations
if Kernel provides it natively. This workaround is not required in Sinatra 2.x