Add pandoc support
Here I try to fix an another warning from tests.
```
cannot load such file -- pandoc-ruby: skipping markdown tests with
Tilt::PandocTemplate
```

`pandoc-ruby` is not defined in Gemfile, so I add it there.
This gem has a prerequisite dependency `pandoc` application that can be
installed via homebrew. That's why I update travis.yml to make it
available during a test run.

The final step to make it work is to exclude `outvar` from the option's
list provided to the template. When `pandoc` application is called
with unsupported parameters it stops the execution and returns an
error.