Install `libyaml-dev` in CI
Needed to be able to install psych >=5.
Need to do it before setup-ruby runs as that runs `bundle install`.

Related to:
- https://github.com/ruby/psych/pull/541
- https://github.com/ruby/setup-ruby/issues/409
- https://github.com/actions/runner-images/issues/6725

Yes, libyaml-dev will be added to GitHub runner images (this week they
say) but opening this PR just in case anyone encounters failing builds.
I don't think it hurts having this in the repo even after libyaml-dev
have been added to the images.

Added some blank lines to make the workflow easier to read.