feature: add indifferent access support to `except`
Starting in Rails 6, the implementation of `except` was changed
and stopped supporting the indifferent access that
`Sinatra::IndifferntHash` provides – calling `except(:a)` on
`Sinatra::IndifferntHash[a: :a]` would not filter the `a` key.

This commits adds indifferent access support to it so it will filter the
key independently if the key is passed as a String or Symbol.