feature: add indifferent access support to `Hash#except`
It was added in Ruby 3.0 so it would be nice that calling except in a
`Sinatra::IndifferntHash` supported the indifferent access that the latter
provides – right now, calling `except(:a)` on `Sinatra::IndifferntHash[a: :a]`
would not filter the `a` key.

This commits adds indifferent access support to except so that it will exclude
the passed key independently if the key is passed as a `String` or `Symbol`.